requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
# Analysis pipeline (scene + segmentation)
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
Pillow>=10.0.0
transformers>=4.30.0
numpy>=1.24.0
scipy>=1.10.0
